rename math_abs to abs
